chore: Add initial scaffolding for core components

+++ src/main/java/com/travelplanner/dto/EventDto.java
@@ -1,7 +1,6 @@
 package com.travelplanner.dto;
 
 import lombok.Data;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,6 +9,6 @@ public class EventDto {
     private String name;
     private String description;
     private LocalDateTime eventDateTime;
-    private BigDecimal price;
+    private Double price;
     private LocationDto location;
 }
